@@ -1,91 +1,67 @@
-import distutils.ccompiler as ccompiler
 import os
 import re
-import tempfile
-from distutils.errors import CCompilerError, DistutilsPlatformError
-from setuptools import setup, find_packages
+from setuptools import setup, find_namespace_packages
 
 install_requires = [
     "aiohttp-json-rpc==0.11.2",
     "aiohttp==3.4.4",
     "appdirs==1.4.3",
-    "async-timeout==3.0.0",
+    "async-timeout==3.0.1",
     "attrs==18.2.0",
     "chardet==3.0.4",
-    "colorama==0.3.9",
-    "discord.py>=1.0.0a0",
+    "colorama==0.4.1",
     "distro==1.3.0; sys_platform == 'linux'",
     "fuzzywuzzy==0.17.0",
     "idna-ssl==1.1.0",
-    "idna==2.7",
-    "multidict==4.4.2",
-    "python-levenshtein==0.12.0",
+    "idna==2.8",
+    "multidict==4.5.2",
+    "python-levenshtein-wheels==0.13.1",
     "pyyaml==3.13",
-    "raven==6.9.0",
+    "raven==6.10.0",
     "raven-aiohttp==0.7.0",
+    "red-lavalink==0.2.0",
     "schema==0.6.8",
     "websockets==6.0",
-    "yarl==1.2.6",
+    "yarl==1.3.0",
 ]
 
 extras_require = {
     "test": [
         "atomicwrites==1.2.1",
-        "more-itertools==4.3.0",
-        "pluggy==0.7.1",
-        "py==1.6.0",
-        "pytest==3.8.2",
-        "pytest-asyncio==0.9.0",
-        "six==1.11.0",
+        "more-itertools==5.0.0",
+        "pluggy==0.8.1",
+        "py==1.7.0",
+        "pytest==4.1.0",
+        "pytest-asyncio==0.10.0",
+        "six==1.12.0",
     ],
-    "mongo": ["motor==2.0.0", "pymongo==3.7.1", "dnspython==1.15.0"],
+    "mongo": ["motor==2.0.0", "pymongo==3.7.2", "dnspython==1.16.0"],
     "docs": [
-        "alabaster==0.7.11",
+        "alabaster==0.7.12",
         "babel==2.6.0",
-        "certifi==2018.8.24",
+        "certifi==2018.11.29",
         "docutils==0.14",
         "imagesize==1.1.0",
         "Jinja2==2.10",
-        "MarkupSafe==1.0",
+        "MarkupSafe==1.1.0",
         "packaging==18.0",
-        "pyparsing==2.2.2",
-        "Pygments==2.2.0",
-        "pytz==2018.5",
-        "requests==2.19.1",
-        "urllib3==1.23",
-        "six==1.11.0",
+        "pyparsing==2.3.0",
+        "Pygments==2.3.1",
+        "pytz==2018.9",
+        "requests==2.21.0",
+        "six==1.12.0",
         "snowballstemmer==1.2.1",
-        "sphinx==1.7.9",
-        "sphinx_rtd_theme==0.4.1",
+        "sphinx==1.8.3",
+        "sphinx_rtd_theme==0.4.2",
         "sphinxcontrib-asyncio==0.2.0",
         "sphinxcontrib-websupport==1.1.0",
+        "urllib3==1.24.1",
     ],
-    "voice": ["red-lavalink==0.1.2"],
-    "style": ["black==18.9b0", "click==7.0", "toml==0.9.6"],
+    "voice": [],
+    "style": ["black==18.9b0", "click==7.0", "toml==0.10.0"],
 }
 
-python_requires = ">=3.6.2,<3.8"
-if os.name == "nt":
-    # Due to issues with ProactorEventLoop prior to 3.6.6 (bpo-26819)
-    python_requires = ">=3.6.6,<3.8"
-
-
-def get_dependency_links():
-    with open("dependency_links.txt") as file:
-        return file.read().splitlines()
-
-
-def check_compiler_available():
-    m = ccompiler.new_compiler()
-
-    with tempfile.TemporaryDirectory() as tdir:
-        with open(os.path.join(tdir, "dummy.c"), "w") as tfile:
-            tfile.write("int main(int argc, char** argv) {return 0;}")
-        try:
-            m.compile([tfile.name], output_dir=tdir)
-        except (CCompilerError, DistutilsPlatformError):
-            return False
-    return True
+python_requires = ">=3.7,<3.8"
 
 
 def get_version():
