@@ -62,7 +62,7 @@ class CoreLogic:
 
     async def _load(
         self, cog_names: Iterable[str]
-    ) -> Tuple[List[str], List[str], List[str], List[str]]:
+    ) -> Tuple[List[str], List[str], List[str], List[str], List[Tuple[str, str]]]:
         """
         Loads cogs by name.
         Parameters
@@ -468,7 +468,7 @@ class Core(commands.Cog, CoreLogic):
 
         pred = MessagePredicate.yes_or_no(ctx)
         try:
-            await self.bot.wait_for("message", check=MessagePredicate.yes_or_no(ctx))
+            await self.bot.wait_for("message", check=pred)
         except asyncio.TimeoutError:
             await ctx.send("Response timed out.")
             return
@@ -500,7 +500,10 @@ class Core(commands.Cog, CoreLogic):
         try:
             await self.bot.wait_for("message", check=pred, timeout=15)
         except asyncio.TimeoutError:
-            await query.delete()
+            try:
+                await query.delete()
+            except discord.errors.NotFound:
+                pass
         else:
             await self.leave_confirmation(guilds[pred.result], ctx)
 
@@ -526,6 +529,8 @@ class Core(commands.Cog, CoreLogic):
     @checks.is_owner()
     async def load(self, ctx: commands.Context, *cogs: str):
         """Loads packages"""
+        if not cogs:
+            return await ctx.send_help()
         async with ctx.typing():
             loaded, failed, not_found, already_loaded, failed_with_reason = await self._load(cogs)
 
@@ -566,6 +571,8 @@ class Core(commands.Cog, CoreLogic):
     @checks.is_owner()
     async def unload(self, ctx: commands.Context, *cogs: str):
         """Unloads packages"""
+        if not cogs:
+            return await ctx.send_help()
         unloaded, failed = await self._unload(cogs)
 
         if unloaded:
@@ -582,6 +589,8 @@ class Core(commands.Cog, CoreLogic):
     @checks.is_owner()
     async def reload(self, ctx: commands.Context, *cogs: str):
         """Reloads packages"""
+        if not cogs:
+            return await ctx.send_help()
         async with ctx.typing():
             loaded, failed, not_found, already_loaded, failed_with_reason = await self._reload(
                 cogs
@@ -1173,6 +1182,9 @@ class Core(commands.Cog, CoreLogic):
             await ctx.send(
                 _("A backup has been made of this instance. It is at {}.").format(backup_file)
             )
+            if backup_file.stat().st_size > 8_000_000:
+                await ctx.send(_("This backup is to large to send via DM."))
+                return
             await ctx.send(_("Would you like to receive a copy via DM? (y/n)"))
 
             pred = MessagePredicate.yes_or_no(ctx)
@@ -1183,10 +1195,18 @@ class Core(commands.Cog, CoreLogic):
             else:
                 if pred.result is True:
                     await ctx.send(_("OK, it's on its way!"))
-                    async with ctx.author.typing():
-                        await ctx.author.send(
-                            _("Here's a copy of the backup"), file=discord.File(str(backup_file))
+                    try:
+                        async with ctx.author.typing():
+                            await ctx.author.send(
+                                _("Here's a copy of the backup"),
+                                file=discord.File(str(backup_file)),
+                            )
+                    except discord.Forbidden:
+                        await ctx.send(
+                            _("I don't seem to be able to DM you. Do you have closed DMs?")
                         )
+                    except discord.HTTPException:
+                        await ctx.send(_("I could not send the backup file."))
                 else:
                     await ctx.send(_("OK then."))
         else:
@@ -1729,7 +1749,7 @@ class Core(commands.Cog, CoreLogic):
         await ctx.tick()
 
     @commands.guild_only()
-    @checks.guildowner_or_permissions(manage_server=True)
+    @checks.guildowner_or_permissions(manage_guild=True)
     @commands.group(name="autoimmune")
     async def autoimmune_group(self, ctx: commands.Context):
         """
